+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-- Added skin pack icons to the change skin screen (Legacy4J only)
-- Improved the arm animations for skins using `animationArmsOutFront`
-- Added single leg animation support for skins using `animationSingleLegAnimation` 
+- Player armor now matches Bedrock player model animations with accurate pivot alignment.
+- Fixed skin customization toggles to save immediately in the skin selection screen.
+- Legacy4J: Player previews in the Change Skin screen now have the walking animation and punching pose.
 
 **Full Changelog**: https://github.com/BrandonItaly/Bedrock-Skins/commits/master